+++ crates/genie-rec/Cargo.toml
@@ -22,6 +22,10 @@ genie-support = { version = "1.0.0", path = "../genie-support", features = [
 ] }
 thiserror.workspace = true
 serde_json = "1.0.64"
+nom = "7.1.1"
+nom-derive = "0.10.0"
+bincode = { version = "1.3.3", features = ["i128"] }
+serde = { version = "1.0.145", features = ["derive", "alloc"] }
 
 [dev-dependencies]
 anyhow.workspace = true
