+++ crates/genie-rec/src/lib.rs
@@ -24,8 +24,11 @@
 pub mod actions;
 pub mod ai;
 pub mod element;
+pub mod error;
+pub mod game_options;
 pub mod header;
 pub mod map;
+pub mod parser;
 pub mod player;
 pub mod reader;
 pub mod string_table;
@@ -36,8 +39,10 @@ pub mod version;
 
 use crate::actions::{Action, Meta};
 use crate::element::ReadableElement;
+use crate::error::Error;
+use crate::error::SyncError;
+use crate::game_options::Difficulty::{Easiest, Extreme, Hard, Hardest, Moderate, Standard};
 use crate::reader::RecordingHeaderReader;
-use crate::Difficulty::{Easiest, Extreme, Hard, Hardest, Moderate, Standard};
 use byteorder::{ReadBytesExt, LE};
 use flate2::bufread::DeflateDecoder;
 use genie_scx::DLCOptions;
@@ -108,49 +113,6 @@ fallible_try_from!(ObjectID, i16);
 fallible_try_from!(ObjectID, i32);
 fallible_try_into!(ObjectID, i16);
 fallible_try_into!(ObjectID, i32);
-
-///
-#[derive(Debug, thiserror::Error)]
-pub enum SyncError {
-    #[error("Got a sync message, but the log header said there would be a sync message {0} ticks later. The recorded game file may be corrupt")]
-    UnexpectedSync(u32),
-    #[error("Expected a sync message at this point, the recorded game file may be corrupt")]
-    ExpectedSync,
-}
-
-/// Errors that may occur while reading a recorded game file.
-#[derive(Debug, thiserror::Error)]
-pub enum Error {
-    #[error(transparent)]
-    IoError(#[from] io::Error),
-    #[error(transparent)]
-    SyncError(#[from] SyncError),
-    #[error(transparent)]
-    DecodeStringError(#[from] genie_support::DecodeStringError),
-    #[error("Could not read embedded scenario data: {0}")]
-    ReadScenarioError(#[from] genie_scx::Error),
-    #[error("Failed to parse DE JSON chat message: {0}")]
-    DEChatMessageJsonError(#[from] serde_json::Error),
-    #[error(
-        "Failed to parse DE JSON chat message, JSON is missing the key {0}, or value is invalid"
-    )]
-    ParseDEChatMessageError(&'static str),
-    #[error(
-    "Failed to find static marker in recording (expected {1:#x} ({1}), found {2:#x} ({2}), version {0}, {3}:{4}, found next {1:#x} ({1}) {5} bytes further)"
-    )]
-    MissingMarker(f32, u128, u128, &'static str, u32, u64),
-    #[error("Failed parsing header at position {0}: {1}")]
-    HeaderError(u64, Box<Error>),
-}
-
-impl From<genie_support::ReadStringError> for Error {
-    fn from(err: genie_support::ReadStringError) -> Self {
-        match err {
-            genie_support::ReadStringError::DecodeStringError(inner) => inner.into(),
-            genie_support::ReadStringError::IoError(inner) => inner.into(),
-        }
-    }
-}
 
 /// Result type alias with `genie_rec::Error` as the error type.
 pub type Result<T> = std::result::Result<T, Error>;
@@ -235,229 +197,6 @@ where
             Err(err) => Some(Err(err.into())),
         }
     }
-}
-
-#[derive(Debug, Clone, Copy, PartialEq, Eq, PartialOrd, Ord, Hash)]
-pub enum StartingResources {
-    None = -1,
-    Standard = 0,
-    Low = 1,
-    Medium = 2,
-    High = 3,
-    UltraHigh = 4,
-    Infinite = 5,
-    Random = 6,
-}
-
-impl From<i32> for StartingResources {
-    fn from(val: i32) -> Self {
-        match val {
-            -1 => StartingResources::None,
-            0 => StartingResources::Standard,
-            1 => StartingResources::Low,
-            2 => StartingResources::Medium,
-            3 => StartingResources::High,
-            4 => StartingResources::UltraHigh,
-            5 => StartingResources::Infinite,
-            6 => StartingResources::Random,
-            _ => unimplemented!("Don't know any starting resource with value {}", val),
-        }
-    }
-}
-
-impl Default for StartingResources {
-    fn default() -> Self {
-        StartingResources::Standard
-    }
-}
-
-#[derive(Debug, Clone, Copy, PartialEq, Eq, PartialOrd, Ord, Hash)]
-pub enum VictoryType {
-    Standard = 0,
-    Conquest = 1,
-    Exploration = 2,
-    Ruins = 3,
-    Artifacts = 4,
-    Discoveries = 5,
-    Gold = 6,
-    TimeLimit = 7,
-    Score = 8,
-    Standard2 = 9,
-    Regicide = 10,
-    LastManStanding = 11,
-}
-
-impl From<u32> for VictoryType {
-    fn from(val: u32) -> Self {
-        match val {
-            0 => VictoryType::Standard,
-            1 => VictoryType::Conquest,
-            2 => VictoryType::Exploration,
-            3 => VictoryType::Ruins,
-            4 => VictoryType::Artifacts,
-            5 => VictoryType::Discoveries,
-            6 => VictoryType::Gold,
-            7 => VictoryType::TimeLimit,
-            8 => VictoryType::Score,
-            9 => VictoryType::Standard2,
-            10 => VictoryType::Regicide,
-            11 => VictoryType::LastManStanding,
-            _ => unimplemented!("Don't know any victory type with value {}", val),
-        }
-    }
-}
-
-impl Default for VictoryType {
-    fn default() -> Self {
-        VictoryType::Standard
-    }
-}
-
-#[derive(Debug, Clone, Copy, PartialEq, Eq, PartialOrd, Ord, Hash)]
-pub enum Difficulty {
-    Easiest = 4,
-    Easy = -1,
-    /// Age of Empires 2: Definitive Edition only.
-    Extreme = 5,
-    Hard = 1,
-    Hardest = 0,
-    Moderate = 2,
-    Standard = 3,
-}
-
-impl From<u32> for Difficulty {
-    fn from(val: u32) -> Self {
-        match val {
-            0 => Hardest,
-            1 => Hard,
-            2 => Moderate,
-            3 => Standard,
-            4 => Easiest,
-            5 => Extreme,
-            _ => unimplemented!("Don't know any difficulty with value {}", val),
-        }
-    }
-}
-
-impl Default for Difficulty {
-    fn default() -> Self {
-        Difficulty::Standard
-    }
-}
-
-#[derive(Debug, Clone, Copy, PartialEq, Eq, PartialOrd, Ord, Hash)]
-pub enum MapSize {
-    Tiny,
-    Small,
-    Medium,
-    Normal,
-    Large,
-    Giant,
-    Ludicrous,
-}
-
-#[derive(Debug, Clone, Copy, PartialEq, Eq, PartialOrd, Ord, Hash)]
-pub enum MapType {}
-
-#[derive(Debug, Clone, Copy, PartialEq, Eq, PartialOrd, Ord, Hash)]
-pub enum ResourceLevel {}
-
-#[derive(Debug, Clone, Copy, PartialEq, Eq, PartialOrd, Ord, Hash)]
-pub enum Age {}
-
-#[derive(Debug, Clone, Copy, PartialEq, Eq, PartialOrd, Ord, Hash)]
-pub enum Visibility {
-    Normal,
-    Explored,
-    AllVisible,
-}
-
-#[derive(Debug, Clone, Copy, PartialEq, Eq, PartialOrd, Ord, Hash)]
-pub enum GameMode {
-    RM = 0,
-    Regicide = 1,
-    DM = 2,
-    Scenario = 3,
-    Campaign = 4,
-    KingOfTheHill = 5,
-    WonderRace = 6,
-    DefendTheWonder = 7,
-    TurboRandom = 8,
-    CaptureTheRelic = 10,
-    SuddenDeath = 11,
-    BattleRoyale = 12,
-    EmpireWars = 13,
-}
-
-impl From<u32> for GameMode {
-    fn from(n: u32) -> Self {
-        match n {
-            0 => GameMode::RM,
-            1 => GameMode::Regicide,
-            2 => GameMode::DM,
-            3 => GameMode::Scenario,
-            4 => GameMode::Campaign,
-            5 => GameMode::KingOfTheHill,
-            6 => GameMode::WonderRace,
-            7 => GameMode::DefendTheWonder,
-            8 => GameMode::TurboRandom,
-            10 => GameMode::CaptureTheRelic,
-            11 => GameMode::SuddenDeath,
-            12 => GameMode::BattleRoyale,
-            13 => GameMode::EmpireWars,
-            _ => unimplemented!("Don't know any game mode with value {}", n),
-        }
-    }
-}
-
-impl Default for GameMode {
-    fn default() -> Self {
-        GameMode::RM
-    }
-}
-
-#[derive(Debug, Clone, Copy, PartialEq, Eq, PartialOrd, Ord, Hash)]
-pub enum GameSpeed {
-    Slow,
-    Casual,
-    Normal,
-    Fast,
-}
-
-#[derive(Debug, Clone)]
-pub struct HDGameOptions {
-    pub dlc_options: DLCOptions,
-    pub difficulty: Difficulty,
-    pub map_size: MapSize,
-    pub map_type: MapType,
-    pub visibility: Visibility,
-    pub starting_resources: ResourceLevel,
-    pub starting_age: Age,
-    pub ending_age: Age,
-    pub game_mode: GameMode,
-    // if version < 1001
-    pub random_map_name: Option<String>,
-    // if version < 1001
-    pub scenario_name: Option<String>,
-    pub game_speed: GameSpeed,
-    pub treaty_length: i32,
-    pub population_limit: i32,
-    pub num_players: i32,
-    pub victory_amount: i32,
-    pub trading_enabled: bool,
-    pub team_bonuses_enabled: bool,
-    pub randomize_positions_enabled: bool,
-    pub full_tech_tree_enabled: bool,
-    pub num_starting_units: i8,
-    pub teams_locked: bool,
-    pub speed_locked: bool,
-    pub multiplayer: bool,
-    pub cheats_enabled: bool,
-    pub record_game: bool,
-    pub animals_enabled: bool,
-    pub predators_enabled: bool,
-    // if version > 1.16 && version < 1002
-    pub scenario_player_indices: Vec<i32>,
 }
 
 /// Recorded game reader.
